requirements.txt
# Base requirements

arrow==0.5.0
boto==2.36.0
celery==3.1.17
Django==1.7.4
django-cors-headers==1.0.0
django-filter==0.9.2
djangorestframework==3.0.4
dj-static==0.0.6
gunicorn==19.2.1
psycopg2==2.5.4
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2

# Testing requirements
coverage==3.7.1
model-mommy==1.2.3
responses==0.3.0
